requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
frozenlist==1.5.0
gTTS==2.5.4
idna==3.10
multidict==6.1.0
openai==0.28.0
playsound==1.2.2
propcache==0.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dotenv==1.0.1
qasync==0.27.1
requests==2.32.3
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
